protect against missing or misconfigured JSON encoder
If the JSON encoder specified by the :json_encoder option (or Poison
by default) does not exist, airbrake_client will compile with a warning.

Also, when Airbrake.Worker starts up, it will fail if the JSON encoder
module is not defined.

Finally, if the module is defined, but it does not define encode!/1,
a warning is printed to stderr AND NO AIRBRAKE NOTICE IS SENT.
Previously the missing-function error would cause the process and the
app to crash.